Db fix

+++ app/Http/Controllers/Admin/DepartmentController.php
@@ -8,6 +8,7 @@ use App\Http\Requests\UpdateDepartmentRequest;
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Validator;
+use Illuminate\Support\Str;
 
 class DepartmentController extends Controller
 {
@@ -52,14 +53,33 @@ class DepartmentController extends Controller
      */
     public function store(StoreDepartmentRequest $request)
     {
-        $validated_data = Validator::make($request->all(), $request->rules(), $request->messages())->validate();
+        $validated_data = Validator::make($request->all(), [
+            'name' => 'required|string|max:255',
+            'description' => 'nullable|string|max:255',
+            'status' => 'nullable',
+        ], [
+            'name.required' => 'Department name is required',
+            'name.string' => 'Department name must be a string',
+            'name.max' => 'Department name must not be greater than 255 characters',
+            'description.string' => 'Department description must be a string',
+            'description.max' => 'Department description must not be greater than 255 characters',
+            'status.required' => 'Department status is required',
+        ]);
 
-        if($validated_data){
-            $department = Department::create($validated_data);
-            return redirect()->back()->with('success', 'Department created successfully');
+        if($validated_data->fails()){
+            dd($validated_data->errors(), $validated_data->messages(), $validated_data->failed());
+            return redirect()->back()->withErrors($validated_data->errors());
         }
+
+        $department                 = new Department();
+        $department->name           = $request->name;
+        $department->description    = $request->description;
+        $department->status         = $request->status;
+        $department->slug           = Str::slug($request->name);
+        $department->save();
         
-        return redirect()->back()->with('error', 'Department creation failed');
+        return redirect()->back()->with('success', 'Department created successfully');
+        
     }
 
     /**
@@ -93,14 +113,18 @@ class DepartmentController extends Controller
      */
     public function update(UpdateDepartmentRequest $request, Department $department)
     {
-        $validated_data = Validator::make($request->all(), $request->rules(), $request->messages())->validate();
+        $validated_data = Validator::make($request->all(), $request->rules(), $request->messages());
 
-        if($validated_data){
-            $department->update($validated_data);
-            return redirect()->back()->with('success', 'Department updated successfully');
+        if($validated_data->fails()){
+            $department->name           = $request->name;
+            $department->description    = $request->description;
+            $department->status         = $request->status;
+            $department->slug           = Str::slug($request->name);
+            $department->update();
+            return redirect()->back()->with(['success' => 'Department updated successfully', 'status' => 200]);
         }
         
-        return redirect()->back()->with('error', 'Department updation failed');
+        return redirect()->back()->with(['error' => 'Department not Updated', 'status' => 404]);
     }
 
     /**
@@ -109,10 +133,17 @@ class DepartmentController extends Controller
      * @param  \App\Models\Department  $department
      * @return \Illuminate\Http\Response
      */
-    public function destroy(Request $id)
+    public function destroy($id)
     {
-        $department = Department::find($id->id);
-        $department->delete();
-        return response()->json(['success' => 'Department deleted successfully', 'status' => 200]);
+        $has_data = Department::find($id)->employees()->exists();
+        
+        if($has_data){
+            return response()->json(['error' => 'Department has employees', 'status' => 400]);
+        }
+        else{
+            $department = Department::find($id);
+            $department->delete();
+            return response()->json(['success' => 'Department deleted successfully', 'status' => 200]);
+        }
     }
 }
